pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Timur <81976283+mym1chelle@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
python-dotenv = "^0.21.0"
dj-database-url = "^1.2.0"
psycopg2 = "^2.9.5"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.3"
whitenoise = "^6.2.0"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
